authctl: Move root command to separate package
So that it can be imported by the docgen tool we're about to add.
