+++ cmd/authctl/root/root.go
@@ -1,0 +1,36 @@
+// Package root contains the root command for authctl.
+package root
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/ubuntu/authd/cmd/authctl/group"
+	"github.com/ubuntu/authd/cmd/authctl/user"
+)
+
+// RootCmd is the root command for authctl.
+var RootCmd = &cobra.Command{
+	Use:   "authctl",
+	Short: "CLI tool to interact with authd",
+	Long:  "authctl is a command-line tool to interact with the authd service for user and group management.",
+	PersistentPreRun: func(cmd *cobra.Command, args []string) {
+		// The command was successfully parsed, so we don't want cobra to print usage information on error.
+		cmd.SilenceUsage = true
+	},
+	CompletionOptions: cobra.CompletionOptions{
+		HiddenDefaultCmd: true,
+	},
+	// We handle errors ourselves
+	SilenceErrors: true,
+	Args:          cobra.NoArgs,
+	RunE:          func(cmd *cobra.Command, args []string) error { return cmd.Usage() },
+}
+
+func init() {
+	// Disable command sorting by name. This makes cobra print the commands in the
+	// order they are added to the root command and adds the `help` and `completion`
+	// commands at the end.
+	cobra.EnableCommandSorting = false
+
+	RootCmd.AddCommand(user.UserCmd)
+	RootCmd.AddCommand(group.GroupCmd)
+}
